fix(payments-paypal): Rename PayPal library filenames
