+++ libs/payments/paypal/src/index.ts
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-export { PayPalClient } from './lib/client';
+export { PayPalClient } from './lib/paypal.client';
+export { PayPalClientError, PayPalNVPError } from './lib/paypal.error';
 export * from './lib/types';
-export { PayPalClientError, PayPalNVPError } from './lib/error';
 export { nvpToObject, objectToNVP, isIpnMerchPmt } from './lib/util';
